Eliminated compilation warnings.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20850 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csplugincommon/canvas/fontcache.cpp
@@ -174,7 +174,7 @@ csFontCache::KnownFont* csFontCache::GetCachedFont (iFont* font)
 	  pg = 0;
 	}
       }
-      knownFont->fontSize = font->GetSize ();
+      knownFont->fontSize = (int)font->GetSize ();
       purgeableFonts.Delete (knownFont);
     }
   }
@@ -185,7 +185,7 @@ csFontCache::KnownFont* csFontCache::CacheFont (iFont* font)
 {
   KnownFont* knownFont = new KnownFont;
   knownFont->font = font;
-  knownFont->fontSize = font->GetSize ();
+  knownFont->fontSize = (int)font->GetSize ();
 
   knownFonts.InsertSorted (knownFont, KnownFontArrayCompareItems);
